+++ config/routes.rb
@@ -43,6 +43,7 @@ Rails.application.routes.draw do
       end
     end
     resources :partners
+    resources :resources
   end
 
   get "learn", to: redirect("education")
@@ -77,6 +78,8 @@ Rails.application.routes.draw do
     get :landing
     get :partners
     get :partner_photo, path: "partners/:id/photo"
+    get :resources
+    get :resource_photo, path: "resources/:id/photo"
     get :privacy_policy, path: "privacy-policy"
     get :sitemap_xml, path: "sitemap.xml.gz"
     get :team